% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqr-package.R
\docType{package}
\name{ineqr-package}
\alias{ineqr}
\alias{ineqr-package}
\title{ineqr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
