% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality.R
\name{inequality}
\alias{inequality}
\title{Inequality}
\usage{
inequality(data, value, weight = NULL, index = "gini")
}
\arguments{
\item{data}{A (grouped) data frame.}

\item{value}{A variable to measure inequality.}

\item{weight}{A variable of weights. By default, no weighting.}

\item{index}{Inequality indices to be measured (\code{"gini"}, \code{"mld"}, \code{"polar"},
or \code{"theil"}). By default, \code{"gini"}.}
}
\value{
A tibble with class \code{inequality} having columns grouping variables,
\code{.index}, \code{inequality}, \code{.cumulative}.
}
\description{
Inequality
}
