% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqr-package.R
\docType{package}
\name{ineqr-package}
\alias{ineqr}
\alias{ineqr-package}
\title{ineqr: Measures and Plots Inequality}
\description{
Measures inequalities such as the Gini coefficient and plots Lorenz curves from (grouped) data frames.
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
